requirements.txt
PyQt6==6.6.1
pyautogui==0.9.54
Pillow==10.2.0
pyzbar==0.1.9
opencv-python==4.9.0.80
numpy==1.26.4 